requirements.txt
torch
lightning
torchvision
opencv-python
pillow
petastorm
petastorm[torch]